V1.0
this was a poorly optimized and quickly done project to make a last minute Christmas gift for a family friend. I may revisit this project and make some changes for better gui control, better performance, and more options, as well as sliders to customize maximum brightness. for now it is archived.